refactor(packaging): data
