+++ app/src/androidTest/java/com/example/ezvault/DeleteItemTest.java
@@ -22,7 +22,7 @@ import com.example.ezvault.model.Item;
 import com.example.ezvault.utils.ItemBuilder;
 import com.example.ezvault.model.ItemList;
 import com.example.ezvault.model.User;
-import com.example.ezvault.utils.UserManager;
+import com.example.ezvault.data.UserManager;
 import com.example.ezvault.view.MainActivity;
 import com.google.firebase.Timestamp;
 import com.google.firebase.auth.FirebaseAuth;
