+++ app/src/androidTest/java/com/example/ezvault/TagListTest.java
@@ -20,7 +20,7 @@ import androidx.test.filters.LargeTest;
 import com.example.ezvault.model.ItemList;
 import com.example.ezvault.model.Tag;
 import com.example.ezvault.model.User;
-import com.example.ezvault.utils.UserManager;
+import com.example.ezvault.data.UserManager;
 import com.example.ezvault.view.MainActivity;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.FirebaseFirestore;
