+++ app/src/main/java/com/example/ezvault/data/authentication/registration/EmailPasswordRegistrationStrategy.java
@@ -1,6 +1,6 @@
-package com.example.ezvault.authentication.registration;
+package com.example.ezvault.data.authentication.registration;
 
-import com.example.ezvault.database.FirebaseBundle;
+import com.example.ezvault.data.database.FirebaseBundle;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
