+++ app/src/main/java/com/example/ezvault/view/fragment/AddItemFragment.java
@@ -38,10 +38,10 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.example.ezvault.view.adapter.PhotoAdapter;
 import com.example.ezvault.R;
 import com.example.ezvault.camera.GalleryAction;
-import com.example.ezvault.database.FirebaseBundle;
-import com.example.ezvault.database.ImageDAO;
-import com.example.ezvault.database.ItemDAO;
-import com.example.ezvault.database.RawUserDAO;
+import com.example.ezvault.data.database.FirebaseBundle;
+import com.example.ezvault.data.database.ImageDAO;
+import com.example.ezvault.data.database.ItemDAO;
+import com.example.ezvault.data.database.RawUserDAO;
 import com.example.ezvault.model.Image;
 import com.example.ezvault.model.Item;
 import com.example.ezvault.utils.ItemBuilder;
@@ -54,7 +54,7 @@ import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
 import com.example.ezvault.upcAPI;
 import com.example.ezvault.utils.FileUtils;
 import com.example.ezvault.utils.TaskUtils;
-import com.example.ezvault.utils.UserManager;
+import com.example.ezvault.data.UserManager;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.Tasks;
 import com.google.android.material.textfield.TextInputLayout;
