+++ app/src/main/java/com/example/ezvault/view/fragment/ItemsFragment.java
@@ -29,15 +29,15 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.example.ezvault.viewmodel.SelectedItemViewModel;
 import com.example.ezvault.view.adapter.ItemAdapter;
 import com.example.ezvault.R;
-import com.example.ezvault.database.FirebaseBundle;
-import com.example.ezvault.database.ImageDAO;
-import com.example.ezvault.database.ItemDAO;
-import com.example.ezvault.database.RawUserDAO;
+import com.example.ezvault.data.database.FirebaseBundle;
+import com.example.ezvault.data.database.ImageDAO;
+import com.example.ezvault.data.database.ItemDAO;
+import com.example.ezvault.data.database.RawUserDAO;
 import com.example.ezvault.model.Item;
 import com.example.ezvault.model.Tag;
 import com.example.ezvault.model.User;
 import com.example.ezvault.utils.ItemListView;
-import com.example.ezvault.utils.UserManager;
+import com.example.ezvault.data.UserManager;
 import com.example.ezvault.viewmodel.ItemViewModel;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
