+++ app/src/main/java/com/example/ezvault/view/fragment/LoginFragment.java
@@ -16,12 +16,12 @@ import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
 
 import com.example.ezvault.R;
-import com.example.ezvault.authentication.authentication.AuthenticationHandler;
-import com.example.ezvault.authentication.authentication.EmailPasswordAuthenticationStrategy;
-import com.example.ezvault.database.FirebaseBundle;
+import com.example.ezvault.data.authentication.authentication.AuthenticationHandler;
+import com.example.ezvault.data.authentication.authentication.EmailPasswordAuthenticationStrategy;
+import com.example.ezvault.data.database.FirebaseBundle;
 import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
 import com.example.ezvault.utils.FragmentUtils;
-import com.example.ezvault.utils.UserManager;
+import com.example.ezvault.data.UserManager;
 
 import javax.inject.Inject;
