+++ app/src/main/java/com/example/ezvault/view/fragment/NewUserFragment.java
@@ -16,15 +16,15 @@ import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
 
 import com.example.ezvault.R;
-import com.example.ezvault.authentication.registration.EmailPasswordRegistrationStrategy;
-import com.example.ezvault.authentication.registration.RegistrationException;
-import com.example.ezvault.authentication.registration.RegistrationHandler;
-import com.example.ezvault.database.FirebaseBundle;
+import com.example.ezvault.data.authentication.registration.EmailPasswordRegistrationStrategy;
+import com.example.ezvault.data.authentication.registration.RegistrationException;
+import com.example.ezvault.data.authentication.registration.RegistrationHandler;
+import com.example.ezvault.data.database.FirebaseBundle;
 import com.example.ezvault.utils.textwatchers.MirroredTextWatcher;
 import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
 import com.example.ezvault.utils.textwatchers.PasswordWatcher;
 import com.example.ezvault.utils.FragmentUtils;
-import com.example.ezvault.utils.UserManager;
+import com.example.ezvault.data.UserManager;
 
 import javax.inject.Inject;
