+++ app/src/main/java/com/example/ezvault/view/fragment/ViewItemFragment.java
@@ -25,16 +25,16 @@ import android.widget.EditText;
 import com.example.ezvault.R;
 import com.example.ezvault.viewmodel.SelectedItemViewModel;
 import com.example.ezvault.view.adapter.ViewPagerAdapter;
-import com.example.ezvault.database.FirebaseBundle;
-import com.example.ezvault.database.ImageDAO;
-import com.example.ezvault.database.ItemDAO;
-import com.example.ezvault.database.RawUserDAO;
+import com.example.ezvault.data.database.FirebaseBundle;
+import com.example.ezvault.data.database.ImageDAO;
+import com.example.ezvault.data.database.ItemDAO;
+import com.example.ezvault.data.database.RawUserDAO;
 import com.example.ezvault.model.Image;
 import com.example.ezvault.model.Item;
 import com.example.ezvault.model.ItemList;
 import com.example.ezvault.model.User;
 import com.example.ezvault.viewmodel.ItemViewModel;
-import com.example.ezvault.utils.UserManager;
+import com.example.ezvault.data.UserManager;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
