+++ app/src/main/java/com/example/ezvault/viewmodel/ItemViewModel.java
@@ -15,7 +15,7 @@ import com.example.ezvault.utils.ItemListView;
 import com.example.ezvault.utils.IItemFilter;
 import com.example.ezvault.utils.ItemListFilter;
 import com.example.ezvault.utils.MainItemFilter;
-import com.example.ezvault.utils.UserManager;
+import com.example.ezvault.data.UserManager;
 
 import java.util.ArrayList;
 import java.util.List;
